project_3/server/requirements.txt
Django==3.2.2
djangorestframework==3.12.4
paho-mqtt==1.5.1